refactor: easterEgg 로직과 컴포넌트 분리
