+++ frontend/src/pages/Landing/Landing.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import styled from 'styled-components';
 
 import ZipgoBanner from '@/assets/webp/landing_banner.webp';
@@ -8,65 +7,46 @@ import Template from '@/components/@common/Template';
 import FilterBottomSheet from '@/components/Food/FilterBottomSheet/FilterBottomSheet';
 import FoodList from '@/components/Food/FoodList/FoodList';
 import FoodSelectionGuideBanner from '@/components/FoodSelectionGuideBanner/FoodSelectionGuideBanner';
-import { useToast } from '@/context/Toast/ToastContext';
+import useEasterEgg from '@/hooks/@common/useEasterEgg';
 
-const TARGET_NUMBER = 100;
-
-let startTime: Date | undefined;
-let endTime: Date | undefined;
-
-const Landing = () => {
-  const { toast } = useToast();
-  const [dogTouchCount, setDogTouchCount] = useState<number>(1);
-  const isDogAwake = dogTouchCount > TARGET_NUMBER;
-
-  const onTouchDog = () => {
-    if (dogTouchCount === 1) {
-      startTime = new Date();
-      toast.info(`강아지를 ${dogTouchCount}번 쓰다듬었어요.`);
-    } else if (dogTouchCount === TARGET_NUMBER && startTime) {
-      endTime = new Date();
-      toast.success(`${(+endTime - +startTime) / 1000}초 만에 강아지를 깨웠어요!`);
-    } else if (TARGET_NUMBER - dogTouchCount === 10) {
-      toast.warning('강아지를 깨우지 않게 조심하세요!');
-    } else if (dogTouchCount < TARGET_NUMBER) {
-      toast.info(`강아지를 ${dogTouchCount}번 쓰다듬었어요.`);
-    }
-
-    if (dogTouchCount <= TARGET_NUMBER) setDogTouchCount(prev => prev + 1);
-  };
-
-  return (
-    <Template staticHeader={Header}>
-      <FoodSelectionGuideBanner />
-      <Layout>
-        <BannerSection>
-          <BannerText>
-            <TitleContainer>
-              <BannerSubTitle>
-                사료 선택이 어려운
-                <br />
-                초보 집사들을 위해
-              </BannerSubTitle>
-              <BannerTitle>집사의고민</BannerTitle>
-            </TitleContainer>
-          </BannerText>
-          <BannerImg
-            src={isDogAwake ? ZipgoBannerAwake : ZipgoBanner}
-            onClick={onTouchDog}
-            alt="집사의고민 배너 이미지"
-          />
-        </BannerSection>
-        <ListSection>
-          <FilterBottomSheet />
-          <FoodList />
-        </ListSection>
-      </Layout>
-    </Template>
-  );
-};
+const Landing = () => (
+  <Template staticHeader={Header}>
+    <FoodSelectionGuideBanner />
+    <Layout>
+      <BannerSection>
+        <BannerText>
+          <TitleContainer>
+            <BannerSubTitle>
+              사료 선택이 어려운
+              <br />
+              초보 집사들을 위해
+            </BannerSubTitle>
+            <BannerTitle>집사의고민</BannerTitle>
+          </TitleContainer>
+        </BannerText>
+        <EasterEggBanner />
+      </BannerSection>
+      <ListSection>
+        <FilterBottomSheet />
+        <FoodList />
+      </ListSection>
+    </Layout>
+  </Template>
+);
 
 export default Landing;
+
+const EasterEggBanner = () => {
+  const { isDogAwake, onTouchDog } = useEasterEgg();
+
+  return (
+    <BannerImg
+      src={isDogAwake ? ZipgoBannerAwake : ZipgoBanner}
+      onClick={onTouchDog}
+      alt="집사의고민 배너 이미지"
+    />
+  );
+};
 
 const Layout = styled.div`
   width: 100%;
